+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ v26.1.11-dev
 **Added Changes**
 
 - Receive all invite-related gateway events.
+- Added :class:`~dsicord.events.InviteDeleteEvent` event.
 
 
 v26.1.10